website/numed-containerized-main/maleo_be/backendapp/requirements.txt
django==3.2
django-storages
django-ckeditor==6.2.0
boto3==1.16.56
django-rest-passwordreset==1.2.0
django-cors-headers
# pillow
djoser==2.1.0
djangorestframework==3.12.4
django-import-export==2.0.2
Pillow
requests
django-rest-swagger
drf-yasg  #swagger buat django 3 +
psycopg2-binary
django-crispy-forms==1.8.1
whitenoise

# restframework jwt
djangorestframework-simplejwt==4.7.2
gunicorn

aiofiles==24.1.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1